move past consensus code to legacy 2
