+++ consensus/legacy/domain/service/election_policy_test.go
@@ -3,7 +3,7 @@ package service
 import (
 	"testing"
 
-	"github.com/it-chain/it-chain-Engine/consensus/domain/model/parliament"
+	"github.com/it-chain/it-chain-Engine/consensus/legacy/domain/model/parliament"
 	"github.com/stretchr/testify/assert"
 )
